Move array, improve blockarray docs
