+++ examples/scripts/denoise_tv_iso_pgm.py
@@ -39,8 +39,8 @@ from xdesign import SiemensStar, discrete_phantom
 import scico.numpy as snp
 import scico.random
 from scico import functional, linop, loss, operator, plot
-from scico.array import ensure_on_device
 from scico.numpy import BlockArray
+from scico.numpy.util import ensure_on_device
 from scico.optimize.pgm import AcceleratedPGM, RobustLineSearchStepSize
 from scico.typing import JaxArray
 from scico.util import device_info
