+++ scico/optimize/_primaldual.py
@@ -14,12 +14,12 @@ from __future__ import annotations
 from typing import Callable, Optional, Union
 
 import scico.numpy as snp
-from scico.array import ensure_on_device
 from scico.diagnostics import IterationStats
 from scico.functional import Functional
 from scico.linop import LinearOperator
 from scico.numpy import BlockArray
 from scico.numpy.linalg import norm
+from scico.numpy.util import ensure_on_device
 from scico.typing import JaxArray
 from scico.util import Timer
